requirements.txt
Flask==2.0.1
PyJWT==2.1.0
PyMySQL==1.0.2
bcrypt==3.2.0
python-dotenv==0.19.0